ipmi_lan: Accept different privilege than requested
Currently openipmi expects that the privilege level in the
open session response equals the privilege which was requested.
This is wrong, because it is legal to request a lower privilege
level than allowed for a user:

According to IPMI v2.0 spec the RCMP+ Open Session Response
contains the _maximum_ privilege Level allowed for a session.

Furthermore Andy Cress pointed out:
Even if the privilege is detected wrong here, it should not abort, but
proceed.

The bug was found because there seems to be a bug in the HP ILO3 IPMI
implementation where the maximum privilege level returned in the response
is "admin" even if the user is only allowed to login with privilege
"operator".

Both bugs together prevent the ipmilan stonith agent to work with ILO3
and fence with priv="operator".

This patch fixes the behavior of openIPMI and allows the "session open"
response message to contain an different privilege level than requested.
An warning is issued if the privilege level is lower than requested.

With this patch fencing with "priv=operator" works correctly.

Signed-off-by: Arnd Hannemann <arnd@arndnet.de>
Signed-off-by: Corey Minyard <cminyard@mvista.com>