lanserv: Fix IANA responses to have IANA number
The IANA numbers are supposed to be in bytes 2-4 off responses to IANA
commands, but they were not being put there.  Fix this in the main IANA
handling to do this automatically, and modify the command handling to
strip the IANA from the command and add it to a message field.  This
makes handling of IANA commands more natural and allows standard message
response generation to work.