lanserv: Rework power handling
If ipmi_sim failed and then restarted when handling a VM, the
VM would reconnect, but it would lose the PID of the VM, so the
power management no longer worked correctly.

Rework the power management a bit so that it bases the power on
status on if a connection exists if the PID is not available, and
add a way to forceably tell the VM to terminate without delay.

Signed-off-by: Corey Minyard <cminyard@vmista.com>