Don't start the power-off timer if already powered off.
When using the 'startcmd' mode the power-off timer should not be
started if the system already is powered off. Otherwise issuing
a power-off followed by almost immediate power-on to a powered off
system will cause the system to initially power up but suddenly
shut down when the timer is triggered.

Signed-off-by: Janne Huttunen <janne.huttunen@nokia.com>
Signed-off-by: Corey Minyard <cminyard@mvista.com>