Fix telnet based Serial-over-LAN simulation.
Don't set the 'write_enabled' field explicitly to one when pushing
the telnet initialization sequence to the buffer. It will be done
by the set_write_enable() call a bit later at the end of the routine.

If the field is explicitly set to one, the set_write_enable() call
won't notify the lower layer and the corresponding file descriptor
will never be enabled for writing. This causes any input coming from
the SoL connection just to be buffed indefinitely and never sent to
the telnet server.

Signed-off-by: Janne Huttunen <janne.huttunen@nokia.com>
Signed-off-by: Corey Minyard <cminyard@mvista.com>