Rework and improve the selector code to be independent of the OS handler
The selector code is useful by itself, don't require an OS handler
for it, pass in the locking functions needed to perform the
threading operation.

This deprecates the old selector allocator function and adds a threaded
and non-threaded version.

This had pretty far-reaching consequences.  The biggest was that the
code that did SNMP had to be redone to work around the silly netsnmp
fd/timer handling.  There were special hacks in the selector code
for this, they were pulled out and a more generic (and complicated)
implementation done.

The selector was also massively improved to add epoll support and
better handling of just about everything.

Signed-off-by: Corey Minyard <cminyard@mvista.com>