sdrcomp: Always round floating point values
It was pointed out that the use of ceil(), truncation, and
rounding in the float to raw value conversion could result in
incorrect values in some instances.  I don't thing there is an
easy way to figure out how to round the value in every situation.
Instead, just do normal rounding and expect the user to give a
floating point value that will convert reasonably close to the
raw value they want.  Rounding will avoid problems with floating
point imprecision.

Also fix a misnamed parameter.

Reported-by: Bjoern Spruck <bspruck@uni-mainz.de>
Signed-off-by: Corey Minyard <cminyard@mvista.com>