sdrcomp: Add a relative threshold floating point value
For hysteresis, it's convenient to be able to specify a floating
point value, but you have to leave the "b" out of the y = mx + b
equation to make it work.  So add a new conversion type for that,
and add hysteresis parameters.

Partially based on a patch from Bjoern Spruck.

Reported-by: Bjoern Spruck <bspruck@uni-mainz.de>
Signed-off-by: Corey Minyard <cminyard@mvista.com>