lanserv: Rework the power cycle and graceful shutdown code
Remove the forceoff function, the normal power off function is a
force off, and replace it with a graceful shutdown.  This is more
clear and accurate.

Rework the power cycle code to wait until the power is "off"
before turning the power back on, and tie in the proper shutdown
code to the watchdog poweroff and powercycle code.

Signed-off-by: Corey Minyard <cminyard@mvista.com>