entity: Rework entity fully up reporting
Knowing when an entity is fully up is a fairly arduous task,
it turns out, and the old code wasn't correct.  For an entity
to be reported as fully up, the SDRs it came from have to be
completely read, the presense and FRU data fetching must be
completely done, and if the entity presence is detected by
the FRU device, the MC must have been reported active/inactive.

This required some hacks in places to know when the MCs have
been scanned (so we know if the MC is inactive) and when
the SDR processing from a repository is complete.

Signed-off-by: Corey Minyard <cminyard@mvista.com>