Add a FRU update handler with an error
FRU updates can fail for various reasons, but there was no way
to report that to the user, a "changed" op was given to the
handler.

Add a new interface that adds an error op (and passes an error
code) to allow an error to be passed.  Deprecate the old
interface.

Also, switch over all the sample code and such to use the
new interface.

Signed-off-by: Corey Minyard <cminyard@mvista.com>