Fix problems compiling under SmartOS
Lots of little things:

You can't take the address of stdout or stderr.

It doesn't have window resize capability.

inet_ntoa is in an inconvenient place.

You have to link with -lsocketlib to get socket calls.

It doesn't have the Linux IPMI driver (of course).

It doesn't like things name s_addr.  That is defined to something some
where.  So change all the s_addr names to s_addr0.

It doesn't have cfmakeraw().

It's ethernet handling functions are in sys/ethernet.h.

For some reason perl didn't like the way the detection of threads and
Tix was done, so rework it to something that is better, anyway.

There was a bogus use of u_int32_t, convert to the standard uint32_t.

epoll on SmartOS is broken.

Signed-off-by: Corey Minyard <cminyard@mvista.com>