Remove selector_t type
The selector_t type was conflicting with something on SmartOS, and
it's really a bad idea to use _t on types outside of the OS.

Unfortunately, I wasn't as wise as I am now about such things when
I first wrong OpenIPMI, and _t is used all over.  The ipmi_xxx_t
ones are probably mostly ok, though.  os_handler_t is probably
pretty safe, too, but maybe not?  No conflicts yet.

selector_t is used by several things that I found, though.  So
I'll remove it.  Most users shouldn't be using it, anyway, so
hoepfully the effect is limited.  And if they are, they can define
OPENIPMI_DEFINE_SELECTOR_T and it will bring the type back.

Signed-off-by: Corey Minyard <cminyard@mvista.com>