atca: Fix removal/insertion handling
If you removed and re-inserted an ATCA FRU, it would delete the FRU
info, not necessarily delete the entity, then when coming back it
would call the update handlers on the entity, with a pointer to
what had previously been freed.

Instead of the current strategy, don't pass the FRU info to the
callback handlers, have them get it from the entity OEM info, and
re-create the FRU info when the entity becomes present again.

Signed-off-by: Corey Minyard <cminyard@mvista.com>