lanserv: Fix send message response routing
Messages sent from the LAN interface routed over BMC would have the
response come back and trigger the BMC interface's ATN bit, but that
bit would not do anything because the receive queue flag wasn't being
set properly, and the message was pulled out of the queue to be send
as the LAN response.

So, we really needed a queue in the channel so LAN and BMC interfaces
weren't sharing one, and we needed proper flag handling all around the
receive queue.

Signed-off-by: Corey Minyard <cminyard@mvista.com>