swig: Rework swig_cb and swig_cb_val
swig changed how it handled these values, they were typedefs of
pointers and I guess it confused swig, because passing "None" to
them caused it to put in a zero.  So make swig_cb and swig_cb_val
typedefs of the main type, and add the "*" everywhere to make them
pointers.

Signed-off-by: Corey Minyard <cminyard@mvista.com>