Fix the semantics of sel_stop_timer_with_done()
sel_stop_timer_with_done() was calling the done handler from
inside itself, but that's not a good idea because it can result in
locking issues.  Instead, always call the done handler from the main
timer processing loop.

Return an error from that function if the timer was stopped or
if a done handler is pending being called.

Signed-off-by: Corey Minyard <cminyard@mvista.com>